+++ backend/services/search-service/internal/.DS_Store
