+++ backend/services/search-service/internal/models/event.go
