+++ backend/services/search-service/internal/models/itinerary.go
